# public/swaggerOutput.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Communication challenge",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:8102",
  "basePath": "/",
  "tags": [],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/communication/": {
      "post": {
        "tags": [
          "Communication"
        ],
        "description": "Endpoint para adicionar um agendamento de envio de comunicação",
        "parameters": [
          {
            "name": "Communication",
            "in": "body",
            "required": "true",
            "description": "Data e Hora para o envio(MM/dd/YYYY hh:mm:ss), \nDestinatário, \nMensagem a ser entregue, \nformatos: email, sms, push e whatsapp",
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "date": {
                  "type": "string",
                  "example": "12/31/2022 23:59:59"
                },
                "recipient": {
                  "type": "string",
                  "example": "João Paulo josé"
                },
                "message": {
                  "type": "string",
                  "example": "Mensagem qualquer"
                },
                "format": {
                  "type": "string",
                  "example": "email"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/communication/{id}": {
      "get": {
        "tags": [
          "Communication"
        ],
        "description": "Endpoint para consultar o status de um agendamento pelo id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "Communication"
        ],
        "description": "Endpoint para cacnelar um um agendamento pelo id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    }
  },
  "definitions": {}
}